[fix #2476] docs: convert useClickableProps stories file to TS
